add class MiniPatch,include SSDAtPoint and FindPatch

+++ include/ImgProc.h
@@ -10,5 +10,5 @@ public:
 
     static int DetectFASTCorners10(const cv::Mat &srcImage, std::vector<cv::Point2f> &vecKeyPoints, const int &threshold=10);
     static double FindShiTomasiScoreAtPoint(const cv::Mat &image, cv::Point2i ptCenter, unsigned int nHalfBoxSize=3);
-    static bool IsInImageWithBorder(const cv::Mat &image, const cv::Point2i &pt, int border=10);
+    static bool IsInImageWithBorder(const cv::Mat &image, const cv::Point2f &pt, int border=10);
 };
