+++ src/ImgProc.cpp
@@ -65,7 +65,7 @@ double ImgProc::FindShiTomasiScoreAtPoint(const cv::Mat &image, cv::Point2i ptCe
 
 }
 
-bool ImgProc::IsInImageWithBorder(const cv::Mat &image, const cv::Point2i &pt, int border)
+bool ImgProc::IsInImageWithBorder(const cv::Mat &image, const cv::Point2f &pt, int border)
 {
     return pt.x >=border && pt.y >=border && pt.x < image.cols - border && pt.y < image.rows - border;
 }
